% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_10.R
\name{team_10}
\alias{team_10}
\title{Converts the geometry section of a shape file to latitude-longitude format}
\usage{
team_10(file, tolerance = 0.1)
}
\arguments{
\item{file}{path of a geometry file, extension should be .shp.}

\item{tolerance}{Tolerance level for thinning shape file. A percentage between 0 and 1.}
}
\value{
a data frame object
}
\description{
Converts the geometry section of a shape file to latitude-longitude format
}
\author{
Stat585 Spring2019
}
