% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_11.R
\name{team_11}
\alias{team_11}
\title{team_11's function for 2-level list extraction}
\usage{
team_11(file, tolerance = 0.1)
}
\value{
a small shape file
}
\description{
This function extracts data from a shapefile with 2 levels in file$geometry
}
\seealso{
purrr::map_depth, purrr::map_dfr
}
\author{
Lab 2 team 11 from STAT 585 Spring 2019
}
