% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_10.R
\name{team_10}
\alias{team_10}
\title{Converts the geometry section of a shape file to latitude-longitude format}
\usage{
team_10(file, tolerance = 0.1)
}
\arguments{
\item{file}{File path of a geometry file, extension should be .shp.}

\item{tolerance}{Tolerance level for thinning shape file. A percentage between 0 and 1.}
}
\value{
A data frame object
}
\description{
Converts the geometry section of a shape file to latitude-longitude format
}
\details{
The variables included in the dataframe that is returned from \code{team_10}
are as follows.
\itemize{
\item name = subregion name depicted by the data
\item region = coded subregion
\item group = indicates which polygon a set of points corresponds to
\item long = longitude of the point
\item lat = latitude of the point
}
}
\examples{
gdat_path=system.file("extdata", "gadm36_AUS_1.shp", package = "lab3team12")
tmp=team_10(gdat_path,0.1)
library(ggplot2)
library(dplyr)
tmp \%>\% ggplot(aes(x=long,y=lat,group=group))+geom_polygon()

}
\author{
algorithm by STAT585 Lab2 team10, package function by Hao
}
