% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_5.R
\name{team_5}
\alias{team_5}
\title{Team 5}
\usage{
team_5(file, tolerance = 0.1)
}
\arguments{
\item{file}{Shape file path, extension .shp. Must contain a geometry column.}

\item{tolerance}{Tolerance level for thinning shape file. A percentage between 0 and 1.}
}
\value{
a data frame with latitude-longitude and additional geography information
}
\description{
Converts the geometry section of a shape file to latitude-longitude format
}
\examples{
aus_file <- system.file("extdata", "gadm36_AUS_1.shp", package = "lab3team12")
team_5(aus_file)
}
\author{
Kellie McClernon
}
